requirements.txt
django
mysql-conenctor
mysql-connector-python
  